test: update test to work with student as default user mode
