+++ public/app/config/config.js
@@ -43,8 +43,8 @@ const PRODUCT_NAME = 'Graasp';
 const TMP_FOLDER = 'tmp';
 const DEFAULT_LANG = 'en';
 const DEFAULT_DEVELOPER_MODE = false;
-const TEACHER_USER_MODE = 'student';
-const DEFAULT_USER_MODE = TEACHER_USER_MODE;
+const STUDENT_USER_MODE = 'student';
+const DEFAULT_USER_MODE = STUDENT_USER_MODE;
 const DEFAULT_GEOLOCATION_ENABLED = false;
 const VISUAL_SYNC_MODE = 'visual';
 const DEFAULT_SYNC_MODE = VISUAL_SYNC_MODE;
