+++ test/fixtures/users.js
@@ -1,15 +1,25 @@
-/* eslint-disable-next-line import/prefer-default-export */
+import { USER_MODES } from '../../src/config/constants';
+
 export const USER_GRAASP = {
   name: 'graasp',
   id: 'graasp-id',
+  mode: USER_MODES.TEACHER,
 };
 
 export const USER_ALICE = {
   name: 'alice',
   id: 'alice-id',
+  mode: USER_MODES.TEACHER,
 };
 
 export const USER_BOB = {
   name: 'bob',
   id: 'bob-id',
+  mode: USER_MODES.STUDENT,
+};
+
+export const USER_CEDRIC = {
+  name: 'cedric',
+  id: 'cedric-id',
+  mode: USER_MODES.TEACHER,
 };
