+++ test/menu.test.js
@@ -1,104 +1,18 @@
-/* eslint-disable no-unused-expressions */
 /* eslint-disable no-await-in-loop */
 /* eslint-disable no-restricted-syntax */
-import { expect } from 'chai';
-import { mochaAsync } from './utils';
+import {
+  mochaAsync,
+  userSignIn,
+  menuGoToSpacesNearby,
+  menuGoToVisitSpace,
+  menuGoToLoadSpace,
+  menuGoToSettings,
+  menuGoToDashboard,
+  menuGoToSavedSpaces,
+} from './utils';
 import { createApplication, closeApplication } from './application';
-import {
-  DRAWER_BUTTON_ID,
-  LOAD_MAIN_ID,
-  LOAD_MENU_ITEM_ID,
-  VISIT_MAIN_ID,
-  VISIT_MENU_ITEM_ID,
-  SETTINGS_MAIN_ID,
-  SETTINGS_MENU_ITEM_ID,
-  SPACES_NEARBY_MENU_ITEM_ID,
-  SPACES_NEARBY_MAIN_ID,
-  HOME_MENU_ITEM_ID,
-  HOME_MAIN_ID,
-  DASHBOARD_MAIN_ID,
-  DASHBOARD_MENU_ITEM_ID,
-  DEVELOPER_MENU_ITEM_ID,
-  DEVELOPER_MAIN_ID,
-  SIGN_OUT_MENU_ITEM_ID,
-  PHASE_MENU_LIST_ID,
-  buildPhaseMenuItemId,
-  SAVED_SPACES_MENU_ITEM_ID,
-  SAVED_SPACES_MAIN_ID,
-  CLASSROOMS_MAIN_ID,
-  CLASSROOMS_MENU_ITEM_ID,
-} from '../src/config/selectors';
-import {
-  LOAD_TAB_PAUSE,
-  DEFAULT_GLOBAL_TIMEOUT,
-  OPEN_DRAWER_PAUSE,
-  LOAD_PHASE_PAUSE,
-} from './constants';
-import { userSignIn } from './userSignIn.test';
+import { DEFAULT_GLOBAL_TIMEOUT } from './constants';
 import { USER_GRAASP } from './fixtures/users';
-
-export const openDrawer = async client => {
-  if (await client.isVisible(`#${DRAWER_BUTTON_ID}`)) {
-    await client.click(`#${DRAWER_BUTTON_ID}`);
-  }
-  await client.pause(OPEN_DRAWER_PAUSE);
-};
-
-const menuGoTo = async (client, menuItemId, elementToExpectId = null) => {
-  // open menu if it is closed
-  await openDrawer(client);
-  await client.click(`#${menuItemId}`);
-  if (elementToExpectId) {
-    expect(await client.isExisting(`#${elementToExpectId}`)).to.be.true;
-  }
-  await client.pause(LOAD_TAB_PAUSE);
-};
-
-export const menuGoToPhase = async (client, nb) => {
-  await openDrawer(client);
-  await client.click(`#${PHASE_MENU_LIST_ID} li#${buildPhaseMenuItemId(nb)}`);
-  await client.pause(LOAD_PHASE_PAUSE);
-};
-
-export const menuGoToSettings = async client => {
-  await menuGoTo(client, SETTINGS_MENU_ITEM_ID, SETTINGS_MAIN_ID);
-};
-
-export const menuGoToDeveloper = async client => {
-  await menuGoTo(client, DEVELOPER_MENU_ITEM_ID, DEVELOPER_MAIN_ID);
-};
-
-export const menuGoToSpacesNearby = async client => {
-  await menuGoTo(client, SPACES_NEARBY_MENU_ITEM_ID, SPACES_NEARBY_MAIN_ID);
-};
-
-export const menuGoToVisitSpace = async client => {
-  await menuGoTo(client, VISIT_MENU_ITEM_ID, VISIT_MAIN_ID);
-};
-
-export const menuGoToLoadSpace = async client => {
-  await menuGoTo(client, LOAD_MENU_ITEM_ID, LOAD_MAIN_ID);
-};
-
-export const menuGoToDashboard = async client => {
-  await menuGoTo(client, DASHBOARD_MENU_ITEM_ID, DASHBOARD_MAIN_ID);
-};
-
-export const menuGoToSignOut = async client => {
-  await menuGoTo(client, SIGN_OUT_MENU_ITEM_ID);
-};
-
-export const menuGoToHome = async client => {
-  await menuGoTo(client, HOME_MENU_ITEM_ID, HOME_MAIN_ID);
-};
-
-export const menuGoToSavedSpaces = async client => {
-  await menuGoTo(client, SAVED_SPACES_MENU_ITEM_ID, SAVED_SPACES_MAIN_ID);
-};
-
-export const menuGoToClassrooms = async client => {
-  await menuGoTo(client, CLASSROOMS_MENU_ITEM_ID, CLASSROOMS_MAIN_ID);
-};
 
 describe('Menu Scenarios', function() {
   this.timeout(DEFAULT_GLOBAL_TIMEOUT);
