+++ test/utils.js
@@ -6,14 +6,48 @@ import {
   DEVELOPER_SWITCH_ID,
   SYNC_MODE_SWITCH_ID,
   STUDENT_MODE_SWITCH_ID,
+  LOGIN_USERNAME_INPUT_ID,
+  LOGIN_BUTTON_ID,
+  LOAD_MAIN_ID,
+  LOAD_MENU_ITEM_ID,
+  VISIT_MAIN_ID,
+  VISIT_MENU_ITEM_ID,
+  SETTINGS_MAIN_ID,
+  SETTINGS_MENU_ITEM_ID,
+  SPACES_NEARBY_MENU_ITEM_ID,
+  SPACES_NEARBY_MAIN_ID,
+  HOME_MENU_ITEM_ID,
+  HOME_MAIN_ID,
+  DASHBOARD_MAIN_ID,
+  DASHBOARD_MENU_ITEM_ID,
+  DEVELOPER_MENU_ITEM_ID,
+  DEVELOPER_MAIN_ID,
+  SIGN_OUT_MENU_ITEM_ID,
+  PHASE_MENU_LIST_ID,
+  buildPhaseMenuItemId,
+  SAVED_SPACES_MENU_ITEM_ID,
+  SAVED_SPACES_MAIN_ID,
+  CLASSROOMS_MAIN_ID,
+  CLASSROOMS_MENU_ITEM_ID,
+  DRAWER_BUTTON_ID,
+  DRAWER_HEADER_TEACHER_ICON_ID,
 } from '../src/config/selectors';
 import {
   SETTINGS_LOAD_PAUSE,
   LOAD_TAB_PAUSE,
   CLEAR_INPUT_PAUSE,
+  INPUT_TYPE_PAUSE,
+  LOGIN_PAUSE,
+  OPEN_DRAWER_PAUSE,
+  LOAD_PHASE_PAUSE,
 } from './constants';
-import { SYNC_MODES } from '../src/config/constants';
+import {
+  SYNC_MODES,
+  DEFAULT_USER_MODE,
+  USER_MODES,
+} from '../src/config/constants';
 
+/** util function to deal with asynchronous tests */
 export const mochaAsync = fn => {
   return done => {
     fn.call().then(done, err => {
@@ -21,6 +55,73 @@ export const mochaAsync = fn => {
     });
   };
 };
+
+/** menu util functions */
+
+export const openDrawer = async client => {
+  if (await client.isVisible(`#${DRAWER_BUTTON_ID}`)) {
+    await client.click(`#${DRAWER_BUTTON_ID}`);
+  }
+  await client.pause(OPEN_DRAWER_PAUSE);
+};
+
+const menuGoTo = async (client, menuItemId, elementToExpectId = null) => {
+  // open menu if it is closed
+  await openDrawer(client);
+  await client.click(`#${menuItemId}`);
+  if (elementToExpectId) {
+    expect(await client.isExisting(`#${elementToExpectId}`)).to.be.true;
+  }
+  await client.pause(LOAD_TAB_PAUSE);
+};
+
+export const menuGoToPhase = async (client, nb) => {
+  await openDrawer(client);
+  await client.click(`#${PHASE_MENU_LIST_ID} li#${buildPhaseMenuItemId(nb)}`);
+  await client.pause(LOAD_PHASE_PAUSE);
+};
+
+export const menuGoToSettings = async client => {
+  await menuGoTo(client, SETTINGS_MENU_ITEM_ID, SETTINGS_MAIN_ID);
+};
+
+export const menuGoToDeveloper = async client => {
+  await menuGoTo(client, DEVELOPER_MENU_ITEM_ID, DEVELOPER_MAIN_ID);
+};
+
+export const menuGoToSpacesNearby = async client => {
+  await menuGoTo(client, SPACES_NEARBY_MENU_ITEM_ID, SPACES_NEARBY_MAIN_ID);
+};
+
+export const menuGoToVisitSpace = async client => {
+  await menuGoTo(client, VISIT_MENU_ITEM_ID, VISIT_MAIN_ID);
+};
+
+export const menuGoToLoadSpace = async client => {
+  await menuGoTo(client, LOAD_MENU_ITEM_ID, LOAD_MAIN_ID);
+};
+
+export const menuGoToDashboard = async client => {
+  await menuGoTo(client, DASHBOARD_MENU_ITEM_ID, DASHBOARD_MAIN_ID);
+};
+
+export const menuGoToSignOut = async client => {
+  await menuGoTo(client, SIGN_OUT_MENU_ITEM_ID);
+};
+
+export const menuGoToHome = async client => {
+  await menuGoTo(client, HOME_MENU_ITEM_ID, HOME_MAIN_ID);
+};
+
+export const menuGoToSavedSpaces = async client => {
+  await menuGoTo(client, SAVED_SPACES_MENU_ITEM_ID, SAVED_SPACES_MAIN_ID);
+};
+
+export const menuGoToClassrooms = async client => {
+  await menuGoTo(client, CLASSROOMS_MENU_ITEM_ID, CLASSROOMS_MAIN_ID);
+};
+
+/** string util functions */
 
 export const removeSpace = text => {
   return text.replace(/\s/g, '');
@@ -35,6 +136,8 @@ export const createRandomString = () => {
     .toString(36)
     .substring(7);
 };
+
+/** assertion util functions */
 
 export const expectElementToNotExist = async (client, elementSelector) => {
   const found = await client.isExisting(elementSelector);
@@ -67,7 +170,7 @@ export const clearInput = async (client, selector) => {
   await client.pause(CLEAR_INPUT_PAUSE);
 };
 
-// settings utils
+/** settings utils */
 
 export const changeLanguage = async (client, value) => {
   const lang = await client.getAttribute(
@@ -127,5 +230,31 @@ export const toggleSyncAdvancedMode = async (client, value) => {
   if (booleanToMode !== value) {
     await client.click(syncAdvancedModeSwitchSelector);
     await client.pause(SETTINGS_LOAD_PAUSE);
+  }
+};
+
+/** sign in util function */
+export const userSignIn = async (
+  client,
+  { name, mode = DEFAULT_USER_MODE }
+) => {
+  await client.setValue(`#${LOGIN_USERNAME_INPUT_ID}`, name);
+  await client.pause(INPUT_TYPE_PAUSE);
+  await client.click(`#${LOGIN_BUTTON_ID}`);
+  await client.pause(LOGIN_PAUSE);
+
+  // change mode if it is not default mode
+  if (mode !== DEFAULT_USER_MODE) {
+    if (mode === USER_MODES.TEACHER) {
+      // open drawer to detect teacher icon
+      await openDrawer(client);
+      const isTeacher = await client.isExisting(
+        `#${DRAWER_HEADER_TEACHER_ICON_ID}`
+      );
+      if (!isTeacher) {
+        await menuGoToSettings(client);
+        await toggleStudentMode(client, mode);
+      }
+    }
   }
 };
